+++ BppLib.Core/CutX.cs
@@ -47,7 +47,7 @@ namespace BppLib.Core
 		public bool Opt { get; set; } = true ;
 		
 		/// <value>Property <c>Th</c> represents the blade thickness.</value>
-		public int Th { get; set; } = 4 ;
+		public double Th { get; set; } = 4 ;
 
 		/// <value>Property <c>Thr</c> represents if needing to execute a through cuts.</value>
 		public bool Thr { get; set; } = false ;
